Refactor `filterTextOnly` utility: move to `ai.koog.prompt.streaming` package and update imports across modules. Add coroutines dependency to `prompt-model`.
