+++ examples/src/main/kotlin/ai/koog/agents/example/structuredoutput/markdown/MarkdownStreamingDataExample.kt
@@ -4,7 +4,7 @@ import ai.koog.agents.core.agent.AIAgent
 import ai.koog.agents.core.agent.config.AIAgentConfig
 import ai.koog.agents.core.dsl.builder.forwardTo
 import ai.koog.agents.core.dsl.builder.strategy
-import ai.koog.agents.core.utils.filterTextOnly
+import ai.koog.prompt.streaming.filterTextOnly
 import ai.koog.agents.example.ApiKeyService
 import ai.koog.prompt.executor.llms.all.simpleOpenAIExecutor
 import ai.koog.prompt.executor.model.PromptExecutor
