+++ prompt/prompt-model/build.gradle.kts
@@ -13,6 +13,7 @@ kotlin {
         commonMain {
             dependencies {
                 api(project(":prompt:prompt-llm"))
+                api(libs.kotlinx.coroutines.core)
                 api(libs.kotlinx.serialization.json)
                 api(libs.kotlinx.datetime)
                 api(libs.kotlinx.io.core)
